image_classifier/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-aiplatform
